trocando de uint16_t para int, as funcoes de get do angulo inicial e funal da barreira

+++ main/include/system_manager.h
@@ -15,7 +15,7 @@
  */
 void system_manager_init(void);
 
-uint16_t get_start_angle_barrier_status();
-uint16_t get_end_angle_barrier_status();
+int get_start_angle_barrier_status();
+int get_end_angle_barrier_status();
 
 #endif /* MAIN_INCLUDE_SYSTEM_MANAGER_H_ */
